// src/types.ts
export type fontWeightType =
  | 'normal'
  | 'bold'
  | '100'
  | '200'
  | '300'
  | '400'
  | '500'
  | '600'
  | '700'
  | '800'
  | '900';

export type fontSizeType =
  | 'h1'
  | 'h2'
  | 'h3'
  | 'h4'
  | 'h5'
  | 'h6'
  | 's1'
  | 's2'
  | 'c1'
  | 'c2'
  | 'p'
  | 'label'
  | 'subtitle'
  | 'icon';

export type fontColorType =
  | 'default'
  | 'primary'
  | 'secondary'
  | 'info'
  | 'caption'
  | 'warning'
  | 'success'
  | 'error';